Create Draft model.ipynb
First draft of the project, this is a kernel pulled from kaggle. (https://www.kaggle.com/code/willkoehrsen/start-here-a-gentle-introduction)
From there on we can start implementing the the final model for predictions and start analyzing the results.

We will use auto-sklearn and a classic neural network architecture to tackle the problem.

SMOTE will be used to deal with the imbalance.